move code to snapshot/.
