+++ lib/trailblazer/developer.rb
@@ -10,6 +10,8 @@ require "trailblazer/activity"
 require "trailblazer/developer/wtf"
 require "trailblazer/developer/wtf/renderer"
 require "trailblazer/developer/trace/snapshot"
+require "trailblazer/developer/trace/snapshot/value"
+require "trailblazer/developer/trace/snapshot/versions"
 require "trailblazer/developer/trace"
 require "trailblazer/developer/trace/stack"
 require "trailblazer/developer/trace/node"
