+++ lib/trailblazer/developer/trace/stack.rb
@@ -8,7 +8,7 @@ module Trailblazer
       # be named "Trace" :D
       # It is by design coupled to both Snapshot and Ctx::Versions.
       class Stack
-        def initialize(snapshots = [], variable_versions = Snapshot::Ctx::Versions.new)
+        def initialize(snapshots = [], variable_versions = Snapshot::Versions.new)
           @snapshots          = snapshots
           @variable_versions  = variable_versions # DISCUSS: I dislike the coupling here to Stack, but introducting another object comprised of Stack and VariableVersions seems overkill.
         end
